+++ .idea/workspace.xml
@@ -11,12 +11,16 @@
     </configurations>
   </component>
   <component name="ChangeListManager">
-    <list default="true" id="14d435eb-22bc-487c-9ad5-73db3aa152ad" name="Changes" comment="add todo">
+    <list default="true" id="14d435eb-22bc-487c-9ad5-73db3aa152ad" name="Changes" comment="update methods    add constructor">
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/CMakeLists.txt" beforeDir="false" afterPath="$PROJECT_DIR$/CMakeLists.txt" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/OrdersDriver.cpp" beforeDir="false" afterPath="$PROJECT_DIR$/OrdersDriver.cpp" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/OrdersList.cpp" beforeDir="false" afterPath="$PROJECT_DIR$/OrdersList.cpp" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/OrdersList.h" beforeDir="false" afterPath="$PROJECT_DIR$/OrdersList.h" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/OrdersList.cpp" beforeDir="false" afterPath="$PROJECT_DIR$/Orders.cpp" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/OrdersList.h" beforeDir="false" afterPath="$PROJECT_DIR$/Orders.h" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/cmake-build-debug/CMakeCache.txt" beforeDir="false" afterPath="$PROJECT_DIR$/cmake-build-debug/CMakeCache.txt" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/cmake-build-debug/CMakeFiles/rules.ninja" beforeDir="false" afterPath="$PROJECT_DIR$/cmake-build-debug/CMakeFiles/rules.ninja" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/cmake-build-debug/Testing/Temporary/LastTest.log" beforeDir="false" afterPath="$PROJECT_DIR$/cmake-build-debug/Testing/Temporary/LastTest.log" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/cmake-build-debug/build.ninja" beforeDir="false" afterPath="$PROJECT_DIR$/cmake-build-debug/build.ninja" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
@@ -66,6 +70,7 @@
     <property name="cidr.known.project.marker" value="true" />
     <property name="cmake.loaded.for.project" value="true" />
     <property name="settings.editor.selected.configurable" value="CPPToolchains" />
+    <property name="structure.view.defaults.are.configured" value="true" />
   </component>
   <component name="RunManager">
     <configuration name="comp335_p" type="CMakeRunConfiguration" factoryName="Application" REDIRECT_INPUT="false" ELEVATE="false" USE_EXTERNAL_CONSOLE="false" PASS_PARENT_ENVS_2="true" PROJECT_NAME="comp335_p" TARGET_NAME="comp335_p" CONFIG_NAME="Debug" RUN_TARGET_PROJECT_NAME="comp335_p" RUN_TARGET_NAME="comp335_p">
@@ -88,7 +93,8 @@
       <workItem from="1644091382473" duration="1218000" />
       <workItem from="1644187273278" duration="1945000" />
       <workItem from="1644261107320" duration="12582000" />
-      <workItem from="1644356970536" duration="11462000" />
+      <workItem from="1644356970536" duration="12332000" />
+      <workItem from="1644433954485" duration="5721000" />
     </task>
     <task id="LOCAL-00001" summary="Create Orderlist files">
       <created>1643947020179</created>
@@ -125,7 +131,14 @@
       <option name="project" value="LOCAL" />
       <updated>1644292937101</updated>
     </task>
-    <option name="localTasksCounter" value="6" />
+    <task id="LOCAL-00006" summary="update methods    add constructor">
+      <created>1644379500327</created>
+      <option name="number" value="00006" />
+      <option name="presentableId" value="LOCAL-00006" />
+      <option name="project" value="LOCAL" />
+      <updated>1644379500327</updated>
+    </task>
+    <option name="localTasksCounter" value="7" />
     <servers />
   </component>
   <component name="TypeScriptGeneratedFilesManager">
@@ -148,6 +161,7 @@
     <MESSAGE value="update OrdersList.cpp and OrdersList.h" />
     <MESSAGE value="update OrdersList.cpp and OrdersList.h  add driver" />
     <MESSAGE value="add todo" />
-    <option name="LAST_COMMIT_MESSAGE" value="add todo" />
+    <MESSAGE value="update methods    add constructor" />
+    <option name="LAST_COMMIT_MESSAGE" value="update methods    add constructor" />
   </component>
 </project>
