+++ Orders.h
@@ -2,25 +2,23 @@
 // Created by admin on 2022/2/3.
 //
 
-#ifndef COMP335_P_ORDERSLIST_H
-#define COMP335_P_ORDERSLIST_H
+#ifndef COMP335_P_ORDERS_H
+#define COMP335_P_ORDERS_H
 
 #include <string>
 using namespace std;
-class OrdersList {
+class Orders {
 public:
 
-    OrdersList();
+    Orders();
     static void initNode(struct Node *head, string n, bool isValidate, int orderID);
     static void addNode(struct Node *head, string n, bool isValidate, int orderID);
     static void insertFront(struct Node **head, string n,bool isValidate,int orderID);
     static bool deleteNode(struct Node **head, Node *ptrDel);
-    static bool remove(struct Node **head, int orderID);
+    static bool remove(struct Node **head, int targetPlace);
     static void deleteLinkedList(struct Node **node);
     static void display(struct Node *head);
-   /* static void move(struct Node **head, int index,int targetPlace );*/
     static Node* searchNode(struct Node *head, int n);
-    static struct Node* findPreTargetNode(struct Node *head,int n);
     static void move(struct Node **head, int index,int targetPlace );
 };
 
@@ -29,6 +27,5 @@ struct Node {
     bool isValidate;
     int orderID;
     Node *next;
-
 };
-#endif //COMP335_P_ORDERSLIST_H
+#endif //COMP335_P_ORDERS_H
