+++ cmake-build-debug/CMakeCache.txt
@@ -113,7 +113,7 @@ CMAKE_INSTALL_PREFIX:PATH=C:/Program Files (x86)/comp335_p
 CMAKE_LINKER:FILEPATH=G:/MYSYS2/mingw64/bin/ld.exe
 
 //make program
-CMAKE_MAKE_PROGRAM:FILEPATH=H:/CLion 2021.3.3/bin/ninja/win/ninja.exe
+CMAKE_MAKE_PROGRAM:FILEPATH=H:/CLION2~1.3/bin/ninja/win/ninja.exe
 
 //Flags used by the linker during the creation of modules during
 // all build types.
