+++ cmake-build-debug/CMakeFiles/rules.ninja
@@ -51,7 +51,7 @@ rule RERUN_CMAKE
 # Rule for cleaning all built files.
 
 rule CLEAN
-  command = "H:\CLion 2021.3.3\bin\ninja\win\ninja.exe" $FILE_ARG -t clean $TARGETS
+  command = H:\CLION2~1.3\bin\ninja\win\ninja.exe $FILE_ARG -t clean $TARGETS
   description = Cleaning all built files...
 
 
@@ -59,6 +59,6 @@ rule CLEAN
 # Rule for printing all primary targets available.
 
 rule HELP
-  command = "H:\CLion 2021.3.3\bin\ninja\win\ninja.exe" -t targets
+  command = H:\CLION2~1.3\bin\ninja\win\ninja.exe -t targets
   description = All primary targets available:
