+++ cmake-build-debug/Testing/Temporary/LastTest.log
@@ -1,3 +1,3 @@
-Start testing: Feb 08 18:57 ??2?㊣那℅?那㊣??
+Start testing: Feb 09 15:07 ??2?㊣那℅?那㊣??
 ----------------------------------------------------------
-End testing: Feb 08 18:57 ??2?㊣那℅?那㊣??
+End testing: Feb 09 15:07 ??2?㊣那℅?那㊣??
