+++ cmake-build-debug/build.ninja
@@ -66,8 +66,8 @@ build CMakeFiles/comp335_p.dir/main.cpp.obj: CXX_COMPILER__comp335_p_Debug C$:/U
   OBJECT_DIR = CMakeFiles\comp335_p.dir
   OBJECT_FILE_DIR = CMakeFiles\comp335_p.dir
 
-build CMakeFiles/comp335_p.dir/OrdersList.cpp.obj: CXX_COMPILER__comp335_p_Debug C$:/Users/admin/CLionProjects/comp345_p/OrdersList.cpp || cmake_object_order_depends_target_comp335_p
-  DEP_FILE = CMakeFiles\comp335_p.dir\OrdersList.cpp.obj.d
+build CMakeFiles/comp335_p.dir/Orders.cpp.obj: CXX_COMPILER__comp335_p_Debug C$:/Users/admin/CLionProjects/comp345_p/Orders.cpp || cmake_object_order_depends_target_comp335_p
+  DEP_FILE = CMakeFiles\comp335_p.dir\Orders.cpp.obj.d
   FLAGS = -g -std=gnu++14
   OBJECT_DIR = CMakeFiles\comp335_p.dir
   OBJECT_FILE_DIR = CMakeFiles\comp335_p.dir
@@ -86,7 +86,7 @@ build CMakeFiles/comp335_p.dir/OrdersDriver.cpp.obj: CXX_COMPILER__comp335_p_Deb
 #############################################
 # Link the executable comp335_p.exe
 
-build comp335_p.exe: CXX_EXECUTABLE_LINKER__comp335_p_Debug CMakeFiles/comp335_p.dir/main.cpp.obj CMakeFiles/comp335_p.dir/OrdersList.cpp.obj CMakeFiles/comp335_p.dir/OrdersDriver.cpp.obj
+build comp335_p.exe: CXX_EXECUTABLE_LINKER__comp335_p_Debug CMakeFiles/comp335_p.dir/main.cpp.obj CMakeFiles/comp335_p.dir/Orders.cpp.obj CMakeFiles/comp335_p.dir/OrdersDriver.cpp.obj
   FLAGS = -g
   LINK_LIBRARIES = -lkernel32 -luser32 -lgdi32 -lwinspool -lshell32 -lole32 -loleaut32 -luuid -lcomdlg32 -ladvapi32
   OBJECT_DIR = CMakeFiles\comp335_p.dir
